Fix debugging in SensorPlantowerPMS; Also fix loop function
As the loop(MyMessage*) method of Sensor is not virtual, derived sensors cannot override it. So instead of setting a flag at the beginning of loop(..), we have to check in onLoop(...) whether the child argument is the first child and in that case set the flag to prevent multiple sensor readouts for one reading.