Memory Optimization: Add NO_MODULE_POWER_MANAGER to remove all traces of the power manager
Currently, each Child and the NodeManager itself hold a pointer to the power manager, even if no power manager is ever needed or desired. This patch adds a NO_MODULE_POWER_MANAGER flag, which will not compile the power manager and also remove all code and variables related to the power manager. This saves at lease one pointer per child plus one pointer in the NodeManager. So with 10 sensors, at least 22 bytes of SRAM and some flash are saved.