Memory Optimization: Add function to List class to pre-allocate a given size
The List class is a candidate to cause some heap fragmentation as the list grows,
as it will always try to re-allocate a new suitable buffer and release the old.
As the NodeManager holds the sensors in a List, each new sensor causes the list
storage to be released and re-allocated. Similarly, sensors with multiple children
currently re-allocate space for the child pointers with each new child. If the list
of sensors and/or the list of children of a sensor is known beforehand, this patch
provides the List::allocateBlocks function to allocate exactly the required memory
once at the beginning so that no heap fragmentation is caused by the list.

Adviced changes to existing code:
- Change "NodeManager node;" to "NodeManager node(3);" in NodeManager.ino, where the
  argument is the required number of sensors (including battery, configuration, etc.)
- In custom sensor implementation with multiple children, add "children.allocateBlocks(3);"
  in the onBefore() method before any of the children are created.