Allow subclasses to override PowerManager methods
This allows users to change the behavior of the PowerManager by
subclassing it. This is the case for example when more than one VCC pin
must be switched on and off at the same time.
Also the member variables need the protected access modifier if one wants
to access them in the subclass.