Fix debug logs in SensorMQ
This fixes the issue when NODEMANAGER_DEBUG is enabled where at the
start of the calibration was missing a new line and there was a shift of the
parameters in the cal. debug print. Furthermore it fixes also the log issue
of the floating point numbers due to an overflow.